\docType{data}
\name{bullrun}
\alias{bullrun}
\title{The Bull Run river flow data (USGS site 14140000) from 1890-01-01 to 2012-09-30}
\description{
  The Bull Run river flow data (USGS site 14140000) from
  1890-01-01 to 2012-09-30
}
\author{
  Jason Law \email{jason.e.law@gmail.com}
}
\references{
  \url{http://waterdata.usgs.gov/nwis}
}
\keyword{data}
\keyword{datasets}

