\name{flow.duration}
\alias{flow.duration}
\title{Plot a flow duration curve}
\usage{
  flow.duration(x, ylab = "Flow",
    xlab = "Exceedance probability", ...)
}
\arguments{
  \item{x}{a zoo object containing flow data}

  \item{ylab}{a y axis label}

  \item{xlab}{an x axis label}

  \item{...}{additional arguments passed to plot}
}
\description{
  Plots a flow duration curve using ecdf() function
}

