\name{IHA-package}
\alias{IHA-package}
\alias{IHA}
\docType{package}
\title{
Indicators of Hydrologic Alteration
}
\description{
Implements The Nature Conservancy's Indicators of Hydrologic Alteration software in R
}
\details{
The group* commands are the workhorse functions.  
}
\author{
Jason Law <jason.e.law@gmail.com>
}
