\name{read.flow}
\alias{read.flow}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Reads a rdb formatted data file from the USGS web site or from a file. }
\description{
  This function reads hydrology data from the waterdata.usgs.gov site or from a file (that is in the USGS rdb format).
}
\usage{
read.flow(site, start.date, end.date = NULL, gauge = F, file = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{site}{ the USGS site_no }
  \item{start.date}{ a starting date in a format convertible to a date using \code{as.Date} }
  \item{end.date}{ an ending data in a format convertible to a date using \code{as.Date} }
  \item{gauge}{ logical describing whether you want gauge height data in addition to flow}
  \item{file}{ a file name if you are reading from a text file }
}
\value{
  A data frame with date and discharge.  If gauge was TRUE, it also includes gauge height.
  Data qualifiers are included also and their meaning can be looked up on the USGS web site.
}
\references{ http://waterdata.usgs.gov/nwis }
\author{ Jason Law <jason.e.law@gmail.com>}

\examples{
d <- read.flow(14211820, start.date = ISOdate(2008,1,1))
head(d)
}



