\name{water.year}
\alias{water.year}
\title{Return a string giving the water year for a date}
\usage{
  water.year(x)
}
\arguments{
  \item{x}{a date-time object which can be handled by
  lubridate}
}
\description{
  Returns a number specifying the water year (\code{2010})
  for a date.
}

