\docType{data}
\name{bullrun}
\alias{bullrun}
\title{Bull Run river flow data}
\description{
  The Bull Run river flow data (USGS site 14140000) from
  1890-01-01 to 2012-09-30. The beginning of operation of
  dam 2 is evident in the drastic changes to low flows
  after 1959.
}
\author{
  Jason Law \email{jason.e.law@gmail.com}
}
\references{
  \url{http://waterdata.usgs.gov/nwis}
}
\keyword{data}
\keyword{datasets}

